Fixed Flower Manager
Also, instead of reading to a list of Vector3's I cut out the middle and
instantiated the objects straight away from the data. There will be a
delay before the flowers appear on the screen when moving the camera, so
some kind of effect like poofing or a revealing,rolling fog should
happen to make it appear more seamless.